requirements.txt
Django==3.0.5
mysqlclient==1.3.13
django-mysql==2.2.0
djangorestframework
django-filter
Pillow
django-nested-admin
django-richtextfield==1.6